Migration support (#147)
Add the capability to load a snapshot from a previous version of a
story.

A created snapshot got an `can_be_migrated()` function to check if the
snapshot is migratable.
Rule of thumb: directly after choosing a choice, before reading any more
content, the state should be migratable, as long as you are not in a
tunnel.

Best practices also include using a hub knot for snapshots with no local
variables, which is guaranteed to be kept around in future versions of
the story.

Also, once only choices should get an explicit label, or you know that
you will not reorder them, because this would break the tracking.

For more details, feel free to ask or look into
`inkcpp\include\snapshot.h`

Fixes: #112 

TODO:
+ [x] add new interface to python
+ [x] add new interface to c

---------

+++ inkcpp_test/FallbackFunction.cpp
@@ -20,9 +20,9 @@ SCENARIO("run a story with external function and fallback function", "[external 
 		WHEN("bind both external functions")
 		{
 			int  cnt_sqrt = 0;
-			auto fn_sqrt  = [&cnt_sqrt](int x) -> int {
+			auto fn_sqrt  = [&cnt_sqrt](double x) -> double {
         ++cnt_sqrt;
-				return static_cast<int>(sqrt(x));
+        return static_cast<int>(sqrt(x));
 			};
 			int  cnt_greeting = 0;
 			auto fn_greeting  = [&cnt_greeting]() -> const char* {
@@ -49,9 +49,9 @@ SCENARIO("run a story with external function and fallback function", "[external 
 		WHEN("only bind function without fallback")
 		{
 			int  cnt_sqrt = 0;
-			auto fn_sqrt  = [&cnt_sqrt](int x) -> int {
+			auto fn_sqrt  = [&cnt_sqrt](double x) -> double {
         ++cnt_sqrt;
-				return static_cast<int>(sqrt(x));
+        return static_cast<int>(sqrt(x));
 			};
 
 			thread->bind("sqrt", fn_sqrt);
